+++ Heiwase.App.Blazor/Components/Pages/AboutSection/AboutSection.razor
@@ -1,0 +1,50 @@
+﻿<section id="about" class="about-section">
+	<div class="about-section-overlay section-padding">
+		<div class="section-header section-header-large-gap">
+			<h2 class="heading-xl">A Csapatról</h2>
+			<p class="about-intro-text">
+				A Heiwase Karate Szeged több mint 30 éve a szegedi sportélet meghatározó szereplője.
+				A csapatot megalakulása óta komplex célok vezérlik. Szeretnénk mindenki számára elérhetővé tenni,
+				és minél több emberrel megismertetni a karatét, biztosítani tanítványaink számára a kellő hátteret,
+				legyen bármi is a céljuk ezzel a sportággal. Akár csak egy jót kíván mozogni a karatéka, akár a
+				magyar válogatott keretbe készül, mi segítünk neki, hogy megvalósítsa azt, amit maga elé
+				tűzött ki célul. A Heiwase büszke az egyedülálló, összetartó közösségére.
+				Csapattagjaink az élet minden területén segítik egymást, összetartanak, barátságokat
+				kötnek. Ott állunk egymás sikerei és kudarcai mögött, bíztatjuk és vigasztaljuk a másikat.
+				Várunk magunkhoz, ha szeretnél egy igazán jó közösséghez tartozni, és jó hangulatban edzeni!
+			</p>
+		</div>
+
+		<div class="about-grid">
+			<div class="info-bubble">
+				<span class="stat-number">1993</span>
+				<h3>óta HEIWASE</h3>
+				<p>
+					Több évtizedes tapasztalattal rendelkezünk a karate oktatásban. A közel 35 éves
+					működése alatt a csapatunk sok mindenen keresztül ment, és számos tapasztalattal
+					gazdagodott, hogy a legjobbat nyújthassa tanítványainak.
+				</p>
+			</div>
+			<div class="info-bubble">
+				<span class="stat-number">100+</span>
+				<h3>tanítvány</h3>
+				<p>
+					Szeged legnagyobb létszámú karatecsapata, közel 100 fővel képviseli városunkat a nagyvilágban.
+					A nagy létszám ellenére a közösségünk befogadó, segítőkész és összetartó, tanítványaink között
+					pedig több fekete öves mestert, és magyar válogatottként világbajnoki éremszerzőt is magunkénak
+					mondhatunk.
+				</p>
+			</div>
+			<div class="info-bubble">
+				<span class="stat-number">0</span>
+				<h3>km-re Szeged szívétől</h3>
+				<p>
+					A Heiwase tatamival fedett, saját edzőtermében várja a karatézni vágyókat. Dojonk a belvárosban
+					található, közel a tömegközlekedési csomópontokhoz, és a Szegedi Tudományegyetemhez. Heti 14 db
+					edzéssel, 3 korcsoporttal, 4 edzéstípussal, reggeli és esti edzésekkel egyedülálló palettát
+					kínál a harcművészetek szerelmeseinek.
+				</p>
+			</div>
+		</div>
+	</div>
+</section>
