+++ Heiwase.App.Blazor/Components/Pages/AboutSection/AboutSection.razor.css
@@ -1,0 +1,39 @@
+﻿.about-section {
+    background: url('img/team.jpg');
+    background-position: center top;
+    background-size: cover;
+    color: white;
+    position: relative;
+}
+
+.about-section-overlay {
+    background: rgba(0, 0, 0, 0.8);
+    width: 100%;
+}
+
+.info-bubble {
+    background: var(--primary-grad);
+    padding: 30px;
+    border-radius: var(--radius);
+    box-shadow: 0 10px 30px rgba(0,0,0,0.3);
+    border-left: 5px solid #cb6ce6;
+    transition: transform 0.4s, background 0.4s;
+    opacity: 0;
+}
+
+    .info-bubble.visible {
+        animation: fadeInUp 0.8s backwards;
+        opacity: 1;
+    }
+
+        .info-bubble.visible:hover {
+            transform: translateY(-5px);
+        }
+
+.stat-number {
+    font-size: 3rem;
+    font-weight: 800;
+    color: #cb6ce6;
+    display: block;
+    font-family: var(--font-heading);
+}
