+++ Heiwase.App.Blazor/Components/Pages/CoachesSection/CoachesSection.razor
@@ -1,0 +1,37 @@
+﻿<section id="coaches" class="section-padding coaches-section">
+	<h2 class="section-title-white">Edzőink</h2>
+	<div class="about-grid about-grid-centered">
+		<div class="card-common">
+			<img class="card-img" src="img/zoli.png" alt="Stempel Zoltán" />
+			<div class="card-details">
+				<h3 class="card-name">Stempel Zoltán</h3>
+				<span class="card-rank">4. dan, vezetőedző</span>
+				<p>
+					Zoltán sensei 1975-ben született, a karatéval 12 éves korában ismerkedett meg.
+					A Wado-ryu előtt Shotokan karatéval és Iaido-val is foglalkozott. A 4. dan övfokozata
+					mellett sportedzői OKJ képesítéssel is rendelkezik, illetve az IWKU
+					(International Wado Karate Union) nemzetközi instruktora. 33 éves edzői tapasztalattal
+					rendelkezik, több korosztályban és felnőtt edzések terén is. 2 lány büszke édesapja,
+					akik közül mindketten aktívan edzenek és segítik edzői munkáját is.
+				</p>
+			</div>
+		</div>
+		<div class="card-common">
+			<img class="card-img" src="img/bendi.png" alt="Tálas Bendegúz" />
+			<div class="card-details">
+				<h3 class="card-name">Tálas Bendegúz</h3>
+				<span class="card-rank">1. dan, edző</span>
+				<p>
+					Tálas Bendegúz 5 éves korában kezdett el karatézni, azóta pedig számos eredményt mondhat
+					magáénak. A Dél-alföldi Karate Akadémia sportolója WKF magyar válogatottként,
+					WKF Magyar Bajnoki ezüst éremmel,
+					és EKF Európa-bajnoki 5. helyezéssel rendelkezik. OKJ-s sportedzői végzettsége mellett,
+					testnevelő tanár hallgató a Szegedi Tudományegyetemen, emellett a Dél-alföldi Karate Akadémia
+					és a Tisza-Volán SC egyesületeknél szerzett edzői tapasztalatokat. A tradicionális karate
+					területén pedig a Shotokan és a Honto-ryu stílusokban szerzett 1. dan
+					övfokozatot.
+				</p>
+			</div>
+		</div>
+	</div>
+</section>
