+++ Heiwase.App.Blazor/Components/Pages/ContactSection/ContactSection.razor.css
@@ -1,0 +1,173 @@
+﻿.contact-section {
+    background: var(--primary-grad);
+}
+
+.heading-white {
+    color: white;
+}
+
+.contact-header {
+    text-align: center;
+    color: white;
+    margin-bottom: 40px;
+}
+
+.contact-form {
+    max-width: 700px;
+    margin: 0 auto;
+    background: white;
+    padding: 40px;
+    border-radius: var(--radius);
+    box-shadow: 0 10px 40px rgba(0,0,0,0.1);
+}
+
+::deep .form-control {
+    width: 100%;
+    padding: 12px;
+    border: 2px solid #eee;
+    border-radius: 10px;
+    font-family: var(--font-body);
+    margin-bottom: 15px;
+}
+
+    ::deep .form-control:focus {
+        border-color: #3533cd;
+        outline: none;
+    }
+
+.btn-submit {
+    width: 100%;
+}
+
+.form-label {
+    display: block;
+    font-weight: 600;
+    margin-bottom: 5px;
+    color: #333;
+    font-size: 0.9rem;
+}
+
+.form-field-collapsible {
+    max-height: 0;
+    opacity: 0;
+    overflow: hidden;
+    transition: max-height 0.4s ease, opacity 0.3s ease, margin 0.3s ease;
+    margin-bottom: 0;
+}
+
+    .form-field-collapsible.form-field-visible {
+        max-height: 160px;
+        opacity: 1;
+        margin-bottom: 0;
+    }
+
+.training-types-grid {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+    margin-top: 5px;
+    margin-bottom: 5px;
+}
+
+.training-type-option {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    padding: 8px 14px;
+    border: 2px solid #eee;
+    border-radius: 20px;
+    cursor: pointer;
+    font-family: var(--font-body);
+    font-size: 0.9rem;
+    transition: border-color 0.25s ease, background 0.25s ease, color 0.25s ease, opacity 0.3s ease;
+    user-select: none;
+}
+
+    .training-type-option:hover:not(.training-type-disabled) {
+        border-color: #3533cd;
+        background: rgba(53, 51, 205, 0.05);
+    }
+
+    .training-type-option input[type="checkbox"] {
+        accent-color: #3533cd;
+        width: 15px;
+        height: 15px;
+        cursor: inherit;
+        flex-shrink: 0;
+    }
+
+.training-type-checked {
+    border-color: #3533cd;
+    background: rgba(53, 51, 205, 0.1);
+    color: #3533cd;
+    font-weight: 600;
+}
+
+.training-type-disabled {
+    opacity: 0.38;
+    cursor: not-allowed;
+}
+
+::deep .validation-message {
+    display: block;
+    color: #dc3545;
+    font-size: 0.82rem;
+    margin-top: -8px;
+    margin-bottom: 10px;
+    animation: fadeInDown 0.25s ease;
+}
+
+.form-feedback-message {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 8px;
+    border-radius: 10px;
+    font-weight: 600;
+    font-size: 0.95rem;
+    text-align: center;
+    max-height: 0;
+    opacity: 0;
+    overflow: hidden;
+    padding: 0 16px;
+    margin-top: 0;
+    transition: max-height 0.4s ease, opacity 0.3s ease, padding 0.3s ease, margin 0.3s ease;
+}
+
+    .form-feedback-message.form-message-visible {
+        max-height: 80px;
+        opacity: 1;
+        padding: 12px 16px;
+        margin-top: 15px;
+    }
+
+.form-success-message {
+    background: #d4edda;
+    color: #155724;
+    border: 1px solid #c3e6cb;
+}
+
+.form-error-message {
+    background: #f8d7da;
+    color: #721c24;
+    border: 1px solid #f5c6cb;
+}
+
+@keyframes fadeInDown {
+    from {
+        opacity: 0;
+        transform: translateY(-6px);
+    }
+
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+.contact-details {
+    margin-top: 30px;
+    text-align: center;
+    border-top: 1px solid #eee;
+    padding-top: 20px;
+}
