+++ Heiwase.App.Blazor/Components/Pages/HallOfFameSection/HallOfFameSection.razor.css
@@ -1,0 +1,112 @@
+﻿.hall-of-fame-section {
+    background: var(--secondary-grad);
+    color: white;
+}
+
+.text-center {
+    text-align: center;
+}
+
+.section-subtitle-white {
+    text-align: center;
+    font-size: 2.5rem;
+    margin-bottom: 30px;
+    color: white;
+}
+
+.section-subtitle-top-gap {
+    margin-top: 60px;
+}
+
+.slider-wrapper {
+    display: flex;
+    align-items: center;
+    gap: 16px;
+    max-width: 1300px;
+    margin: 0 auto 60px auto;
+    padding: 0 20px;
+}
+
+.slider-container {
+    position: relative;
+    flex: 1;
+    min-width: 0;
+    overflow: hidden;
+}
+
+.slider-nav-btn {
+    flex: 0 0 48px;
+    width: 48px;
+    height: 48px;
+    border-radius: 50%;
+    background: var(--primary-grad);
+    color: white;
+    border: none;
+    cursor: pointer;
+    font-size: 1.2rem;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    transition: transform 0.2s;
+}
+
+    .slider-nav-btn:hover {
+        transform: scale(1.08);
+    }
+
+.slider-grid {
+    display: flex;
+    flex-wrap: nowrap;
+    gap: 30px;
+    will-change: transform;
+}
+
+.slider-grid > .card-common {
+    flex: 0 0 calc((100% - 60px) / 3);
+    min-width: 0;
+}
+
+.slider-controls {
+    display: flex;
+    justify-content: center;
+    gap: 20px;
+    margin-top: 20px;
+}
+
+.slider-btn {
+    background: var(--primary-grad);
+    color: white;
+    border: none;
+    width: 50px;
+    height: 50px;
+    border-radius: 50%;
+    cursor: pointer;
+    font-size: 1.2rem;
+    transition: 0.3s;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+    .slider-btn:hover {
+        transform: scale(1.1);
+        background: var(--secondary-grad);
+    }
+
+@media (max-width: 768px) {
+    .slider-wrapper {
+        gap: 8px;
+        padding: 0 10px;
+    }
+
+    .slider-nav-btn {
+        flex: 0 0 36px;
+        width: 36px;
+        height: 36px;
+        font-size: 1rem;
+    }
+
+    .slider-grid > .card-common {
+        flex: 0 0 100%;
+    }
+}
