+++ Heiwase.App.Blazor/Components/Pages/MwkszSection/MwkszSection.razor
@@ -1,0 +1,70 @@
+﻿<section id="mwksz" class="section-padding mwksz-section">
+	<div class="section-header section-header-medium-gap">
+		<h2 class="heading-xl">Szövetségünk</h2>
+		<p class="mwksz-intro-text">
+			A Heiwase a Magyarországi Wado-ryu Karate Szövetség elkötelezett tagja.
+			Az MWKSZ biztosítja klubunk számára a szakmai hátteret, és validációt,
+			mindemellett a WUKF (World Union of Karate-do Federations) tagjaként lehetőséget
+			biztosít csapatunknak a nemzetközi versenyzésre is, míg
+			az IWKU (International Wado Karate Union) alapító tagjaként
+			képzi tovább csapattagjainkat, Shibamori Kando (9. dan)
+			nemzetközi főinstruktor irányítása alatt. Vezetője Rostás Tibor (7. dan).
+		</p>
+	</div>
+
+	<div class="features-grid">
+		<div class="feature-card">
+			<i class="fas fa-graduation-cap feature-icon"></i>
+			<h4>Övvizsgák</h4>
+			<p>
+				Stempel Zoltán az MWKSZ övvizsgáztatói jogosultságával rendelkezik,
+				így csapatunk jogosult övvizsgákat szervezni. Emellett csapatunk tagjai
+				rendszeresen vesznek részt sikerrel szövetségi övvizsgákon, hogy barna, illetve
+				fekete övvfokozatokat szerzzenek.
+			</p>
+		</div>
+		<div class="feature-card">
+			<i class="fas fa-campground feature-icon"></i>
+			<h4>Edzőtáborok</h4>
+			<p>
+				Shibamori Kando, 9. danos japán nagymester nyaranta hazánkba látogat,
+				hogy ott 3 nemzet karatékáinak tartson nemzetközi edzőtábort, illetve
+				övvizsgáztasson. Ezen egyedülálló lehetőségget a Heiwase karatékái minden
+				évben kihasználják, és együtt edzenek Shibamori mesterrel.
+			</p>
+		</div>
+		<div class="feature-card">
+			<i class="fas fa-trophy feature-icon"></i>
+			<h4>Versenyek</h4>
+			<p>
+				Az MWKSZ Országos Bajnokságán a Heiwase minden évben előkelő helyet foglal el a tabellán.
+				A 2025-ös szezonban 3 db arany, 3 ezüst és 6 bronzéremmel gazdagították versenyzőink a csapatot.
+				A Heiwase jelentős nemzetközi eredményekkel is rendelkezik, jelenleg is több karatékánk
+				a magyar válogatott keret tagja, Zimmermann Zsolt Gábor (1. kyu), pedig a 2023-mas világbajnokságról
+				bronzéremmel tért haza.
+			</p>
+		</div>
+		<div class="feature-card">
+			<i class="fas fa-globe-europe feature-icon"></i>
+			<h4>Szemináriumok</h4>
+			<p>
+				Rostás Tibor (7. dan) hazai stílusvezető, és Shibamori Kando (9. dan) nagymester évente több
+				alkalommal tart szemináriumot és tréninget az MWKSZ karatékáinak. Többször tudtuk sikerként
+				elkönyvelni, hogy a japán nagymester Szegedre is lelátogatott, ahol a Heiwase látta vendégül
+				a 80 éves shihant. A továbbképzéseken mindig nagy számmal képviseljük magunkat, az ott
+				szerzett tudást pedig a legtöbbször övvfokozatokra váltják tanítványaink a vizsgákon.
+			</p>
+		</div>
+	</div>
+
+	<div class="gallery-collage">
+		<img src="img/mwkszgrid/csoport.jpg" class="gallery-img g-item-1" />
+		<img src="img/mwkszgrid/katica.jpg" class="gallery-img g-item-2" />
+		<img src="img/mwkszgrid/technika.jpg" class="gallery-img g-item-3" />
+		<img src="img/mwkszgrid/verseny.jpg" class="gallery-img g-item-4" />
+	</div>
+
+	<div class="event-calendar-wrapper">
+		<img src="img/esemenynaptar26.jpg" class="event-calendar-image" />
+	</div>
+</section>
