+++ Heiwase.App.Blazor/Components/Pages/MwkszSection/MwkszSection.razor.css
@@ -1,0 +1,95 @@
+﻿.mwksz-section {
+    background: var(--primary-grad);
+    color: var(--text-light);
+}
+
+.mwksz-intro-text {
+    max-width: 600px;
+    justify-self: center;
+}
+
+.features-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+    gap: 20px;
+    max-width: 1200px;
+    margin: 2rem auto;
+}
+
+.feature-icon {
+    font-size: 2.5rem;
+    margin-bottom: 15px;
+}
+
+.gallery-collage {
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    grid-template-rows: repeat(2, 250px);
+    gap: 15px;
+    max-width: 1200px;
+    margin: 0 auto 60px auto;
+}
+
+.gallery-img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    border-radius: var(--radius);
+    transition: transform 0.3s;
+    cursor: pointer;
+}
+
+    .gallery-img:hover {
+        transform: scale(1.02);
+        z-index: 2;
+        box-shadow: 0 10px 20px rgba(0,0,0,0.2);
+    }
+
+.g-item-1 {
+    grid-column: span 2;
+    grid-row: span 2;
+}
+
+.g-item-2 {
+    grid-column: span 1;
+    grid-row: span 1;
+}
+
+.g-item-3 {
+    grid-column: span 1;
+    grid-row: span 2;
+}
+
+.g-item-4 {
+    grid-column: span 1;
+    grid-row: span 1;
+}
+
+.event-calendar-wrapper {
+    display: flex;
+    justify-content: center;
+}
+
+.event-calendar-image {
+    border-radius: 50px;
+    justify-self: center;
+    max-width: 60%;
+    height: auto;
+}
+
+@media (max-width: 768px) {
+    .event-calendar-image {
+        max-width: 100%;
+    }
+
+    .gallery-collage {
+        grid-template-columns: 1fr;
+        grid-template-rows: auto;
+    }
+
+    .g-item-1, .g-item-2, .g-item-3, .g-item-4 {
+        grid-column: span 1;
+        grid-row: span 1;
+        height: 250px;
+    }
+}
