+++ Heiwase.App.Blazor/Components/Pages/NavigationMenu/NavigationMenu.razor
@@ -1,0 +1,18 @@
+﻿<nav class="sticky-nav">
+	<img src="img/logo.png" width="50" alt="Logo" />
+
+	<div class="hamburger" @onclick="ToggleMenu">
+		<i class="fas fa-bars"></i>
+	</div>
+
+	<div class="nav-links @(_isMenuOpen ? "active" : "")">
+		<a href="#home" @onclick="CloseMenu">Kezdőlap</a>
+		<a href="#why-us" @onclick="CloseMenu">Edzéstípusok</a>
+		<a href="#about" @onclick="CloseMenu">Rólunk</a>
+		<a href="#coaches" @onclick="CloseMenu">Edzők</a>
+		<a href="#mwksz" @onclick="CloseMenu">Szövetség</a>
+		<a href="#halloffame" @onclick="CloseMenu">Hall of Fame</a>
+		<a href="#timetable" @onclick="CloseMenu">Edzésrend</a>
+		<a href="#contact" @onclick="CloseMenu">Kapcsolat</a>
+	</div>
+</nav>
