+++ Heiwase.App.Blazor/Components/Pages/NavigationMenu/NavigationMenu.razor.cs
@@ -1,0 +1,10 @@
+﻿namespace Heiwase.App.Blazor.Components.Pages.NavigationMenu;
+
+public partial class NavigationMenu
+{
+    private bool _isMenuOpen = false;
+
+    private void ToggleMenu() => _isMenuOpen = !_isMenuOpen;
+
+    private void CloseMenu() => _isMenuOpen = false;
+}
