+++ Heiwase.App.Blazor/Components/Pages/NavigationMenu/NavigationMenu.razor.css
@@ -1,0 +1,67 @@
+﻿.sticky-nav {
+    position: fixed;
+    top: 20px;
+    right: 20px;
+    background: var(--primary-grad);
+    backdrop-filter: blur(10px);
+    padding: 10px 25px;
+    border-radius: 50px;
+    box-shadow: 0 5px 15px rgba(0,0,0,0.3);
+    z-index: 1000;
+    display: flex;
+    align-items: center;
+    gap: 20px;
+    transition: all 0.3s ease;
+}
+
+.nav-links {
+    display: flex;
+    gap: 20px;
+}
+
+.sticky-nav a {
+    text-decoration: none;
+    text-transform: uppercase;
+    color: #ffffff;
+    font-weight: 700;
+    font-size: 1.1rem;
+    transition: 0.3s;
+}
+
+    .sticky-nav a:hover {
+        color: #cb6ce6;
+    }
+
+.hamburger {
+    display: none;
+    cursor: pointer;
+    color: white;
+    font-size: 1.5rem;
+}
+
+@media (max-width: 768px) {
+    .sticky-nav {
+        top: 0;
+        right: 0;
+        width: 100%;
+        border-radius: 0 0 20px 20px;
+        justify-content: space-between;
+        flex-wrap: wrap;
+    }
+
+    .nav-links {
+        display: none;
+        flex-direction: column;
+        width: 100%;
+        text-align: center;
+        padding-bottom: 20px;
+    }
+
+        .nav-links.active {
+            display: flex;
+        }
+
+    .hamburger {
+        display: block;
+    }
+}
