+++ Heiwase.App.Blazor/Components/Pages/SiteFooter/SiteFooter.razor
@@ -1,0 +1,4 @@
+﻿<footer>
+	<p>© @DateTime.Now.Year Heiwase Karate Szeged. Minden jog fenntartva.</p>
+	<p class="footer-credit">Készítette: Zimmermann Zsolt Gábor | .NET Development</p>
+</footer>
