+++ Heiwase.App.Blazor/Components/Pages/SiteFooter/SiteFooter.razor.css
@@ -1,0 +1,10 @@
+﻿footer {
+    background: #111;
+    color: #888;
+    text-align: center;
+    padding: 40px 20px;
+}
+
+.footer-credit {
+    font-size: 0.8rem;
+}
