+++ Heiwase.App.Blazor/Components/Pages/SocialSection/SocialSection.razor
@@ -1,0 +1,31 @@
+﻿<section class="section-padding social-section">
+	<h2 class="section-subtitle-white heading-xl heading-bottom-gap">Kövess minket!</h2>
+	<div class="social-grid">
+		<div class="social-box social-box-surface social-box-facebook">
+			<div class="fb-page social-facebook-embed">
+				<iframe class="social-facebook-iframe" src="https://www.facebook.com/plugins/page.php?href=https%3A%2F%2Fwww.facebook.com%2Fszegedkaratehirek%2F&tabs=timeline&small_header=false&adapt_container_width=true&hide_cover=false&show_facepile=true"
+						scrolling="no"
+						frameborder="0"
+						allow="autoplay; clipboard-write; encrypted-media; picture-in-picture; web-share">
+				</iframe>
+			</div>
+		</div>
+		<div class="social-box social-box-surface">
+			<div>
+				<i class="fab fa-instagram fa-3x social-icon social-icon-instagram"></i>
+				<p>Kövess minket az Instagramon!</p>
+				<a href="https://instagram.com/szegedkarate" target="_blank">@@szegedkarate</a>
+			</div>
+			<div>
+				<i class="fab fa-tiktok fa-3x social-icon social-icon-tiktok"></i>
+				<p>Kövess minket a TikTokon!</p>
+				<a href="" target="_blank">@@szegedkarate</a>
+			</div>
+			<div>
+				<i class="fab fa-youtube fa-3x social-icon social-icon-youtube"></i>
+				<p>Iratkozz fel YouTube csatornánkra!</p>
+				<a href="https://www.youtube.com/channel/UCmw-zZurtwfqpPsgWuwagpw" target="_blank">Heiwase Karate Szeged</a>
+			</div>
+		</div>
+	</div>
+</section>
