+++ Heiwase.App.Blazor/Components/Pages/SocialSection/SocialSection.razor.css
@@ -1,0 +1,88 @@
+﻿.social-section {
+    background: var(--secondary-grad);
+    text-align: center;
+}
+
+.heading-bottom-gap {
+    margin-bottom: 30px;
+}
+
+.social-grid {
+    display: flex;
+    justify-content: center;
+    gap: 50px;
+    flex-wrap: wrap;
+}
+
+.social-box {
+    width: 400px;
+    height: 500px;
+    background: #eee;
+    border-radius: var(--radius);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    color: #888;
+    gap: 20px;
+}
+
+.social-box-facebook {
+    padding: 20px;
+    overflow: hidden;
+}
+
+.social-facebook-embed {
+    width: 100%;
+    height: 100%;
+}
+
+.social-facebook-iframe {
+    display: block;
+    width: 100%;
+    height: 100%;
+    border: 0;
+    overflow: hidden;
+}
+
+.social-box-surface {
+    border: 1px solid #ddd;
+    background: white;
+}
+
+.social-icon {
+    margin-bottom: 10px;
+}
+
+.social-icon-facebook {
+    color: #1877f2;
+    display: block;
+}
+
+.social-icon-instagram {
+    color: #c32aa3;
+}
+
+@media (max-width: 768px) {
+    .social-grid {
+        gap: 20px;
+    }
+
+    .social-box {
+        width: 100%;
+        max-width: 600px;
+        height: auto;
+        min-height: 560px;
+        gap: 24px;
+    }
+
+    .social-box-facebook {
+        padding: 12px;
+        min-height: 0;
+        aspect-ratio: 500 / 650;
+    }
+
+    .social-facebook-embed {
+        min-height: 0;
+    }
+}
