+++ Heiwase.App.Blazor/Components/Pages/StickyCta/StickyCta.razor
@@ -1,0 +1,3 @@
+﻿<a href="#contact" class="sticky-cta" title="Jelentkezés">
+	<i class="fas fa-user-plus"></i> <p class="sticky-cta-text">CSATLAKOZZ MÉG MA!</p>
+</a>
