+++ Heiwase.App.Blazor/Components/Pages/TimetableSection/TimetableSection.razor
@@ -1,0 +1,14 @@
+﻿<section id="timetable" class="timetable-section">
+	<div class="timetable-section-overlay section-padding">
+		<div class="schedule-map-container">
+			<div>
+				<h2 class="timetable-heading">Edzésrend</h2>
+				<img src="img/edzeseink.png" class="timetable-image" alt="Edzésidőpontok" />
+			</div>
+			<div>
+				<h2 class="timetable-heading">Heiwase dojo</h2>
+				<iframe class="map-frame" src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2758.895392576326!2d20.14728507673322!3d46.25203397109848!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x47448865664d42d3%3A0xc008e980f278e41e!2sHeiwase%20wado-ryu%20karate%20klub!5e0!3m2!1shu!2shu!4v1700000000000!5m2!1shu!2shu" allowfullscreen="" loading="lazy"></iframe>
+			</div>
+		</div>
+	</div>
+</section>
