+++ Heiwase.App.Blazor/Components/Pages/TimetableSection/TimetableSection.razor.css
@@ -1,0 +1,40 @@
+﻿.timetable-section {
+    background: url('img/edzes.jpg');
+}
+
+.timetable-section-overlay {
+    background: rgba(0, 0, 0, 0.7);
+}
+
+.schedule-map-container {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 20px;
+    max-width: 1200px;
+    margin: 0 auto;
+}
+
+.timetable-heading {
+    font-size: 2.5rem;
+    margin-bottom: 20px;
+    color: white;
+}
+
+.timetable-image {
+    width: 100%;
+    border-radius: 20px;
+    box-shadow: 0 5px 15px rgba(0,0,0,0.1);
+}
+
+.map-frame {
+    width: 100%;
+    height: 400px;
+    border-radius: var(--radius);
+    border: 0;
+}
+
+@media (max-width: 768px) {
+    .schedule-map-container {
+        grid-template-columns: 1fr;
+    }
+}
