+++ Heiwase.App.Blazor/Components/Pages/WhyUsSection/WhyUsSection.razor
@@ -1,0 +1,55 @@
+﻿<section id="why-us" class="section-padding why-us-section">
+	<div class="section-header section-header-large-gap">
+		<h2 class="heading-xxl">Edzéstípusaink</h2>
+		<p class="about-intro-text lead-text">
+			Változatos edzésprogramunkból mindenki megtalálhatja a számára legmegfelelőbbet.
+			A tradicionális Wado-ryu karate mellett sportkarate, női önvédelem és atlétikai edzések is várják a jelentkezőket.
+			Bármi is a célod – legyen az egészséges életmód, önvédelem vagy profi versenysport –, mi azért dolgozunk,
+			hogy megvalósíthasd azokat csapatunkban. Nem számít, hogy most kezded, vagy tapasztalt versenyző vagy-e,
+			ahogyan az sem, hogy mennyire érzed magad tehetségesnek. Csatlakozz hozzánk bizalommal!
+		</p>
+	</div>
+
+	<div class="why-us-grid" id="animated-cards">
+		<div class="feature-card">
+			<i class="fas fa-torii-gate feature-icon-large"></i>
+			<h3>Wado-ryu</h3>
+			<p>
+				Jelentése magyarul: a béke útjának iskolája. A 4 alap karatestílus egyike, mely Okinawa szigetéről származik.
+				Alapítómestere a megelőzésben és a gyors, kitérő mozdulatokban látta meg a hatékony közdelem alapjait. Edzéseinken megismerkedhetsz
+				a Wado-ryu karate mozdulataival és technikáival, formagyakorlatokat tanulhatsz meg, és páros küzdelmi feladatokat is elsajátíthatsz.
+				Ha pedig szeretnél komolyabban elmélyülni a tanulásban, akkor övvizsgákon adhatsz számot a megszerzett tudásodról.
+			</p>
+		</div>
+		<div class="feature-card feature-card-delay-02">
+			<i class="fas fa-medal feature-icon-large"></i>
+			<h3>Sportkarate</h3>
+			<p>
+				A karatéban fellelhető rengeteg stílus, iskola, szövetség és szabályrendszer egyesítése révén született meg a sportkarate. Ez egy olyan
+				biztonságos, modern mozgásforma, mely semi-contact (azaz visszafogott erősségű) technikákkal dolgozik, a küzdelmek pedig
+				védőfelszerelésekben zajlanak. Ha szeretnéd magad megmérettetni egy karate versenyen, esetleg profi sportolói álmokat dédelgetsz, vagy csak
+				érdekel ez a küzdelmi rendszer, akkor a sportkarate edzéseinket neked találták ki!
+			</p>
+		</div>
+		<div class="feature-card feature-card-delay-04">
+			<i class="fas fa-shield-alt feature-icon-large"></i>
+			<h3>Női önvédelem</h3>
+			<p>
+				A karate elsődleges célja mindig is az önvédelem, a váratlan támadások hatékony hatástalanítása volt. Számos projekt született már arra,
+				hogy a karate technikákat alapul véve kifejezetten a nők számára fejlesszenek önvédelmi rendszereket. Ezt most te is kipróbálhatod!
+				Az edzéseinken megtanulhatod megvédeni magadat olyan esetekben, amikor már semmi más opciód nincs a támadóval szemben.
+				Emellett figyelünk a kellő mennyiségű erősítő és mozgásfejlesztő gyakorlatokra is, hogy a tested és az elméd is felkészült legyen a váratlan helyzetekre.
+			</p>
+		</div>
+		<div class="feature-card feature-card-delay-06">
+			<i class="fas fa-dumbbell feature-icon-large"></i>
+			<h3>Atlétika</h3>
+			<p>
+				Ezen edzéseink olyan plusz alkalmakként szolgálnak, amelyeken a karatéban elengedhetetlen erő, állóképesség, gyorsaság és mozgékonyság
+				fejlesztésére helyezzük a hangsúlyt. Fárasztó, de jó hangulatú edzéseken főleg saját testsúlyos feladatok várnak rád. Az edzéseken
+				részt vehetsz külön is, ha csak egy jót szeretnél mozogni, de a karate edzéseinket is kiegészítheted vele, hogy a lehető legjobb formába hozd magad.
+				Várunk sok szeretettel!
+			</p>
+		</div>
+	</div>
+</section>
