+++ Heiwase.App.Blazor/Components/Pages/WhyUsSection/WhyUsSection.razor.css
@@ -1,0 +1,40 @@
+﻿.why-us-section {
+    background: var(--primary-grad);
+    color: white;
+    overflow: hidden;
+}
+
+.why-us-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+    gap: 40px;
+    max-width: 1300px;
+    margin: 0 auto;
+}
+
+.feature-card-delay-02 {
+    animation-delay: 0.2s;
+}
+
+.feature-card-delay-04 {
+    animation-delay: 0.4s;
+}
+
+.feature-card-delay-06 {
+    animation-delay: 0.6s;
+}
+
+.feature-icon-large {
+    font-size: 4rem;
+    margin-bottom: 20px;
+    color: white;
+    text-shadow: 0 5px 15px rgba(0,0,0,0.2);
+}
+
+.heading-xxl {
+    font-size: 3.5rem;
+}
+
+.lead-text {
+    font-size: 1.2rem;
+}
