+++ Heiwase.App.Blazor/wwwroot/css/app.css
@@ -139,3 +139,139 @@ code {
 .form-floating > .form-control-plaintext:focus::placeholder, .form-floating > .form-control:focus::placeholder {
     text-align: start;
 }
+
+/* Shared home-page section styles reused across multiple page section components. */
+
+.section-padding {
+    padding: 80px 20px;
+}
+
+.text-gradient {
+    background: var(--primary-grad);
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+}
+
+.section-header {
+    text-align: center;
+}
+
+.section-header-large-gap {
+    margin-bottom: 50px;
+}
+
+.section-header-medium-gap {
+    margin-bottom: 40px;
+}
+
+.heading-xl {
+    font-size: 3rem;
+}
+
+.about-intro-text {
+    max-width: 900px;
+    margin: 0 auto;
+    font-size: 1.2rem;
+}
+
+.btn-design {
+    background: var(--primary-grad);
+    color: white;
+    border: none;
+    padding: 15px 30px;
+    border-radius: 30px;
+    font-size: 1.2rem;
+    cursor: pointer;
+    font-weight: bold;
+    transition: transform 0.2s;
+}
+
+    .btn-design:hover {
+        transform: scale(1.02);
+    }
+
+.feature-card {
+    background: rgba(255, 255, 255, 0.1);
+    border: 1px solid rgba(255, 255, 255, 0.2);
+    padding: 40px 20px;
+    border-radius: var(--radius);
+    text-align: center;
+    transition: transform 0.4s, background 0.4s;
+    opacity: 0;
+}
+
+    .feature-card.visible {
+        animation: fadeInUp 0.8s backwards;
+        opacity: 1;
+    }
+
+        .feature-card.visible:hover {
+            transform: translateY(-10px);
+            background: rgba(255, 255, 255, 0.2);
+        }
+
+@keyframes fadeInUp {
+    0% {
+        opacity: 0;
+        transform: translateY(30px);
+    }
+
+    100% {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+.about-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+    gap: 30px;
+    max-width: 1200px;
+    margin: 0 auto;
+}
+
+.about-grid-centered {
+    justify-content: center;
+}
+
+.card-common {
+    background: var(--primary-grad);
+    border-radius: var(--radius);
+    overflow: hidden;
+    box-shadow: 0 10px 20px rgba(0,0,0,0.1);
+    text-align: center;
+    border-bottom: 5px solid #3533cd;
+    color: var(--text-light);
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    transition: transform 0.3s;
+}
+
+.card-img {
+    width: 100%;
+    height: 300px;
+    object-fit: cover;
+    background-color: #ddd;
+}
+
+.card-details {
+    padding: 20px;
+    flex-grow: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+}
+
+.card-name {
+    font-size: 2rem;
+    margin: 0;
+    color: white;
+}
+
+.card-rank {
+    color: whitesmoke;
+    font-weight: 600;
+    margin-bottom: 10px;
+    display: block;
+}
